pyproject.toml
[project]
name = "substar"
version = "0.1.0"
description = "A social media app"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-fasthtml>=0.4.5",
    "supabase>=2.7.3",
    "python-dotenv>=1.0.1",
]
